+++ backend/src/main/java/site/kkokkio/global/exception/doc/ErrorCode.java
@@ -11,6 +11,7 @@ public enum ErrorCode {
 	REFRESH_TOKEN_NOT_FOUND("401-1", "리프레시 토큰이 없습니다."),
 	REFRESH_TOKEN_MISMATCH("401-2", "유효하지 않은 리프레시 토큰입니다."),
 	REFRESH_TOKEN_INTERNAL_ERROR("500", "리프레시 토큰 처리 중 서버 오류가 발생했습니다."),
+	MEMBER_ME_BAD_REQUEST("400", "마이페이지 조회 실패"),
 	COMMENT_NOT_FOUND("404", "존재하지 않는 댓글입니다."),
 	COMMENT_UPDATE_FORBIDDEN("403", "본인 댓글만 수정할 수 있습니다."),
 	COMMENT_DELETE_FORBIDDEN("403", "본인 댓글만 삭제할 수 있습니다."),
