+++ backend/src/test/java/site/kkokkio/domain/member/controller/MemberControllerV1Test.java
@@ -19,17 +19,14 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import site.kkokkio.domain.member.controller.dto.MemberLoginResponse;
 import site.kkokkio.domain.member.controller.dto.MemberResponse;
 import site.kkokkio.domain.member.controller.dto.MemberSignUpRequest;
-import site.kkokkio.domain.member.dto.TokenDto;
 import site.kkokkio.domain.member.service.AuthService;
 import site.kkokkio.domain.member.service.MailService;
 import site.kkokkio.domain.member.service.MemberService;
 import site.kkokkio.global.aspect.ResponseAspect;
 import site.kkokkio.global.enums.MemberRole;
-import site.kkokkio.global.exception.ServiceException;
+import site.kkokkio.global.exception.CustomAuthException;
 import site.kkokkio.global.util.JwtUtils;
 
 @WebMvcTest(MemberControllerV1.class)
@@ -77,7 +74,7 @@ class MemberControllerV1Test {
 			""";
 
 		// when & then
-		mockMvc.perform(post("/api/v1/auth/signup")
+		mockMvc.perform(post("/api/v1/member/signup")
 				.contentType(MediaType.APPLICATION_JSON)
 				.content(signupJson))
 			.andExpect(status().isOk())
@@ -90,170 +87,46 @@ class MemberControllerV1Test {
 	}
 
 	@Test
-	@DisplayName("로그인 - 성공")
-	void loginSuccess() throws Exception {
+	@DisplayName("회원 정보 조회 - 성공")
+	void getMember_success() throws Exception {
 		// given
-		MemberLoginResponse loginResponse = MemberLoginResponse.builder()
-			.nickname("테스트쟁이")
-			.email("test@test.com")
-			.deleteAt(null)
-			.role(MemberRole.USER)
-			.token("tokenTest")
-			.build();
-		given(authService.login(eq("test@test.com"), eq("passHash"), any(HttpServletResponse.class)))
-			.willReturn(loginResponse);
-		willDoNothing().given(jwtUtils)
-			.setJwtInCookie(eq("tokenTest"), any(HttpServletResponse.class));
-
-		String loginJson = """
-			{
-			  "email":"test@test.com",
-			  "passwordHash":"passHash"
-			}
-			""";
+		MemberResponse respDto = new MemberResponse(
+			"user@example.com",
+			"tester",
+			LocalDate.of(1990, 1, 1),
+			MemberRole.USER
+		);
+		given(memberService.getMemberInfo(any(HttpServletRequest.class)))
+			.willReturn(respDto);
 
 		// when & then
-		mockMvc.perform(post("/api/v1/auth/login")
-				.contentType(MediaType.APPLICATION_JSON)
-				.content(loginJson))
+		mockMvc.perform(get("/api/v1/member/me")
+				.accept(MediaType.APPLICATION_JSON))
 			.andExpect(status().isOk())
 			.andExpect(jsonPath("$.code").value("200"))
-			.andExpect(jsonPath("$.message").value("로그인 성공"))
-			.andExpect(jsonPath("$.data.nickname").value("테스트쟁이"))
-			.andExpect(jsonPath("$.data.email").value("test@test.com"))
+			.andExpect(jsonPath("$.message").value("마이페이지 조회 성공"))
+			.andExpect(jsonPath("$.data.email").value("user@example.com"))
+			.andExpect(jsonPath("$.data.nickname").value("tester"))
+			.andExpect(jsonPath("$.data.birthDate").value("1990-01-01"))
 			.andExpect(jsonPath("$.data.role").value("USER"));
 	}
 
 	@Test
-	@DisplayName("로그인 - 실패 (인증 오류)")
-	void loginFailure() throws Exception {
-		// given
-		given(authService.login(anyString(), anyString(), any(HttpServletResponse.class)))
-			.willThrow(new ServiceException("401", "로그인에 실패했습니다."));
-
-		String loginJson = """
-			{
-			  "email":"bad@test.com",
-			  "passwordHash":"wrong"
-			}
-			""";
+	@DisplayName("회원 정보 조회 - 인증 토큰 누락으로 실패")
+	void getMember_noToken_fail() throws Exception {
+		// given: 토큰 누락 시 CustomAuthException 발생
+		given(memberService.getMemberInfo(any(HttpServletRequest.class)))
+			.willThrow(new CustomAuthException(
+				CustomAuthException.AuthErrorType.CREDENTIALS_MISMATCH,
+				"인증 토큰이 없습니다."
+			));
 
 		// when & then
-		mockMvc.perform(post("/api/v1/auth/login")
-				.contentType(MediaType.APPLICATION_JSON)
-				.content(loginJson))
+		mockMvc.perform(get("/api/v1/member/me")
+				.accept(MediaType.APPLICATION_JSON))
 			.andExpect(status().isUnauthorized())
-			.andExpect(jsonPath("$.code").value("401"))
-			.andExpect(jsonPath("$.message").value("로그인에 실패했습니다."));
+			.andExpect(jsonPath("$.status_code").value(401))
+			.andExpect(jsonPath("$.err_code").value("CREDENTIALS_MISMATCH"))
+			.andExpect(jsonPath("$.message").value("인증 토큰이 없습니다."));
 	}
-
-	@Test
-	@DisplayName("토큰 재발급 - 성공")
-	void refreshTokenSuccess() throws Exception {
-		// given
-		TokenDto tokenDto = new TokenDto("newAccessToken", "existingRefreshToken");
-
-		given(authService.refreshToken(any(HttpServletRequest.class), any(HttpServletResponse.class)))
-			.willReturn(tokenDto);
-
-		// when & then
-		mockMvc.perform(post("/api/v1/auth/refresh"))
-			.andExpect(status().isOk())
-			.andExpect(jsonPath("$.code").value("200"))
-			.andExpect(jsonPath("$.message").value("토큰이 재발급되었습니다."))
-			.andExpect(jsonPath("$.data.accessToken").value("newAccessToken"))
-			.andExpect(jsonPath("$.data.refreshToken").value("existingRefreshToken"));
-	}
-
-	@Test
-	@DisplayName("로그아웃 - 성공")
-	void logoutSuccess() throws Exception {
-		// given
-		willDoNothing().given(authService)
-			.logout(any(HttpServletRequest.class), any(HttpServletResponse.class));
-
-		// when & then
-		mockMvc.perform(post("/api/v1/auth/logout"))
-			.andExpect(status().isOk())
-			.andExpect(jsonPath("$.code").value("200"))
-			.andExpect(jsonPath("$.message").value("로그아웃 되었습니다."));
-	}
-
-	@Test
-	@DisplayName("이메일 인증 코드 전송 - 성공")
-	void sendAuthCodeSuccess() throws Exception {
-		// given
-		given(mailService.sendAuthCode("test@example.com"))
-			.willReturn(true);
-
-		// when & then
-		mockMvc.perform(post("/api/v1/auth/verify-email")
-				.param("email", "test@example.com"))
-			.andExpect(status().isOk())
-			.andExpect(jsonPath("$.code").value("200"))
-			.andExpect(jsonPath("$.message").value("인증 코드가 전송되었습니다."));
-	}
-
-	@Test
-	@DisplayName("이메일 인증 코드 전송 - 실패")
-	void sendAuthCodeFailure() throws Exception {
-		// given
-		given(mailService.sendAuthCode("test@example.com"))
-			.willReturn(false);
-
-		// when & then
-		mockMvc.perform(post("/api/v1/auth/verify-email")
-				.param("email", "test@example.com"))
-			.andExpect(status().isInternalServerError())
-			.andExpect(jsonPath("$.code").value("500"))
-			.andExpect(jsonPath("$.message").value("인증 코드 전송이 실패하였습니다."));
-	}
-
-	// 이메일 인증 검사는 실제 메일로 검사 필요
-
-	// @Test
-	// @DisplayName("이메일 인증 검사 - 성공")
-	// void checkEmailSuccess() throws Exception {
-	// 	// given
-	// 	given(mailService.validationAuthCode(any(EmailVerificationRequest.class)))
-	// 		.willReturn(true);
-	//
-	// 	String checkJson = """
-	// 		{
-	// 		  "email":"test@example.com",
-	// 		  "code":"123456"
-	// 		}
-	// 		""";
-	//
-	// 	// when & then
-	// 	mockMvc.perform(post("/api/v1/auth/check-email")
-	// 			.contentType(MediaType.APPLICATION_JSON)
-	// 			.content(checkJson))
-	// 		.andExpect(status().isOk())
-	// 		.andExpect(jsonPath("$.code").value("200"))
-	// 		.andExpect(jsonPath("$.message").value("이메일 인증에 성공하였습니다."));
-	// }
-	//
-	// @Test
-	// @DisplayName("이메일 인증 검사 - 실패")
-	// void checkEmailFailure() throws Exception {
-	// 	// given
-	// 	given(mailService.validationAuthCode(any(EmailVerificationRequest.class)))
-	// 		.willReturn(false);
-	//
-	// 	String checkJson = """
-	// 		{
-	// 		  "email":"test@example.com",
-	// 		  "code":"000000"
-	// 		}
-	// 		""";
-	//
-	// 	// when & then
-	// 	mockMvc.perform(post("/api/v1/auth/check-email")
-	// 			.contentType(MediaType.APPLICATION_JSON)
-	// 			.content(checkJson))
-	// 		.andExpect(status().isBadRequest())
-	// 		.andExpect(jsonPath("$.code").value("400"))
-	// 		.andExpect(jsonPath("$.message").value("이메일 인증에 실패하였습니다."));
-	// }
 }
